+++ src/features/accounts/components/account-link.tsx
@@ -1,20 +1,13 @@
 import { PropsWithChildren } from 'react'
-import { useLoadableNfd } from '@/features/nfd/data/nfd'
+import { useLoadableNfdResult } from '@/features/nfd/data/nfd'
 import { RenderLoadable } from '@/features/common/components/render-loadable'
-import { is404 } from '@/utils/error'
-import { AccountLinkInner } from './account-link-inner'
-
-const nfdInvalidAddressMessage = 'Invalid nfd address'
-const nfdFailedToLoadMessage = 'Nfd failed to load'
-
-export const transformError = (e: Error) => {
-  if (is404(e)) {
-    return new Error(nfdInvalidAddressMessage)
-  }
-  // eslint-disable-next-line no-console
-  console.error(e)
-  return new Error(nfdFailedToLoadMessage)
-}
+import { fixedForwardRef } from '@/utils/fixed-forward-ref'
+import { CopyButton } from '@/features/common/components/copy-button'
+import { cn } from '@/features/common/utils'
+import { useSelectedNetwork } from '@/features/network/data'
+import { TemplatedNavLink } from '@/features/routing/components/templated-nav-link/templated-nav-link'
+import { Urls } from '@/routes/urls'
+import { ellipseAddress } from '@/utils/ellipse-address'
 
 export type AccountLinkProps = PropsWithChildren<{
   address: string
@@ -24,13 +17,59 @@ export type AccountLinkProps = PropsWithChildren<{
 }>
 
 export const AccountLink = ({ address, ...rest }: AccountLinkProps) => {
-  const [loadablenfd] = useLoadableNfd(address)
+  const [loadableNfd] = useLoadableNfdResult(address)
 
   return (
     <>
-      <RenderLoadable loadable={loadablenfd} transformError={transformError} fallback={<AccountLinkInner address={address} {...rest} />}>
+      <RenderLoadable loadable={loadableNfd} fallback={<AccountLinkInner address={address} {...rest} />}>
         {(nfd) => <AccountLinkInner address={address} nfd={nfd?.name ?? undefined} {...rest} />}
       </RenderLoadable>
     </>
   )
 }
+
+type AccountLinkInnerProps = AccountLinkProps & {
+  nfd?: string
+}
+
+const AccountLinkInner = fixedForwardRef(
+  (
+    { address, nfd, short, className, children, showCopyButton, ...rest }: AccountLinkInnerProps,
+    ref?: React.LegacyRef<HTMLAnchorElement>
+  ) => {
+    const [selectedNetwork] = useSelectedNetwork()
+
+    const link = (
+      <TemplatedNavLink
+        className={cn(!children && 'text-primary underline', !children && !short && 'truncate', className)}
+        urlTemplate={Urls.Explore.Account.ByAddress}
+        urlParams={{ address, networkId: selectedNetwork }}
+        ref={ref}
+        {...rest}
+      >
+        {children ? (
+          children
+        ) : nfd ? (
+          <abbr className="tracking-wide" title={address}>
+            {nfd}
+          </abbr>
+        ) : short ? (
+          <abbr className="tracking-wide" title={address}>
+            {ellipseAddress(address)}
+          </abbr>
+        ) : (
+          address
+        )}
+      </TemplatedNavLink>
+    )
+
+    return children ? (
+      link
+    ) : (
+      <div className="flex items-center">
+        {link}
+        {showCopyButton && <CopyButton value={address} />}
+      </div>
+    )
+  }
+)
