+++ src/features/accounts/components/labels.ts
@@ -20,7 +20,7 @@ export const accountOptedApplicationsTabId = 'opted-applications'
 
 export const accountInformationLabel = 'Account Information'
 export const accountAddressLabel = 'Address'
-export const accountNfdLabel = 'Nfd'
+export const accountNfdLabel = 'NFD'
 export const accountBalanceLabel = 'Balance'
 export const accountMinBalanceLabel = 'Min Balance'
 export const accountAssetsHeldLabel = 'Held assets'
